+++ src/components/Form/TextareaField.js
@@ -1,4 +1,4 @@
-function TextareaForm(name) {
+function TextareaField(name) {
   const textareaElement = document.createElement("textarea");
 
   textareaElement.name = name;
@@ -9,4 +9,4 @@ function TextareaForm(name) {
   return textareaElement;
 }
 
-export default TextareaForm;
+export default TextareaField;
